Implemented the fork optimization to reduce shim overhead.
Previously, the shim executed the module load, and then immediately
executed the lambda handler function, all within a single sandbox
execution.

Now, the sandbox executes the module load (i.e, runs the original code
of the lambda). This creates a sandboxed js module with handler
funtions that can be executed. The shim then creates handler functions
that perform a fork in which the child process analyzes the event
object, populates the label, and executes the handler function of the
sandboxed module.

In theory, this should drastically cut down the shim execution
overhead. The shim overhead should now consist mainly of event
processing (stripping labels) and execution label assignment, which is
about the bare minimum, and cannot be avoided in any case, with the
addition of the OS overhead involved in forking the process. By
comparison, the original shim overhead also included the creation of
the vm and the loading of the lambda module (and any depencies
therein) in every lambda invocation (minus the fork overhead).