A working version of the fork optimization.
The authentication service caused problems with the fork --- the authmap promise was resolved after the fork, creating a duplicate resolution of the same mysql server session, which in turn led to an error when the parent process attempted to communicate with the server (after the child process had already closed the connection with the server).
To fix this, I added a `fence' method to the authentication module, and call the fence in the handler method, ensuring that everithing (in particular, the fork) is executed after the authentication module has resolved the authentication map.